Internal methods

FiniteHorizonGramians._exp_and_gram_chol_init — Method

_exp_and_gram_init(A::AbstractMatrix{T}, B::AbstractMatrix{T}, L::LegendreExp{T})

Computes the matrix exponential exp(A) and the controllability Grammian ∫_0^1 exp(At)BB'exp(A'*t) dt, using a Legendre expansion of the matrix exponential.

FiniteHorizonGramians._dims_if_compatible — Method

_dims_if_compatible(A::AbstractMatrix, B::AbstractMatrix)

Throws DimensionMismatch if A is not square or if the number of rows of B does not equal the number of columns of A. Is equivalent to size(B) if no error is thrown.

FiniteHorizonGramians.triu2cholesky_factor! — Method

triu2cholesky_factor!(A::AbstractMatrix{T})

If A is an upper triangular matrix, it computes the product QA in-place, where Q is a unitary transform such that QA is a valid Cholesky factor. If A is not an upper triangular matrix, returns garbage.

FiniteHorizonGramians._symmetrize! — Method

_symmetrize!(A::AbstractMatrix{T}) where {T<:Number}

Discards the skew-Hermitian part of A in-place.