Algorithm types

FiniteHorizonGramians.AdaptiveExpAndGram — Type

AdaptiveExpAndGram{T,A} <: AbstractExpAndGramAlgorithm

Adaptive algorithm for computing the matrix exponential and an associated Gramian.

Constructor:

AdaptiveExpAndGram{T}()

creates an algorithm with coefficients stored in the numeric type T.

FiniteHorizonGramians.ExpAndGram — Type

ExpAndGram{T,N,A,B,C} <: AbstractExpAndGramAlgorithm

Non-adaptive algorithm of order N for computing the matrix exponential and an associated Gramian.

Constructor:

ExpAndGram{T,N}()

creates an algorithm with coefficients stored in the numeric type T of order N. Current supported values of N are 3, 5, 7, 9, 13.

Methods

FiniteHorizonGramians.exp_and_gram — Function

exp_and_gram(A::AbstractMatrix, B::AbstractMatrix, alg::AbstractExpAndGramAlgorithm)

Compute the matrix exponential exp(A) and the controllability Gramian of (A, B) over the unit interval.

FiniteHorizonGramians.exp_and_gram_chol — Function

exp_and_gram_chol(A::AbstractMatrix, B::AbstractMatrix, alg::AbstractExpAndGramAlgorithm)

Compute the matrix exponential exp(A) and a upper triangular Cholesky factor of the controllability Gramian of (A, B) over the unit interval.